// Web/RMA-Web-Shell/projects/clientcare/src/app/broker-manager/models/organisation-option-item-value.ts
export class OrganisationOptionItemValue {
    organisationOptionItemValueId: number;
	benefitOptionItemValueId: number;
	productOptionOptionItemValueId: number;
	brokerageId: number;
	rolePlayerId: number;
	effectiveDate: Date;
	value: number;
	isExtended: boolean;
	isDeleted: boolean;
	createdBy: string;
	createdDate: Date;
	modifiedBy: string;
	modifiedDate: Date;
}