// Web/RMA-Web-Shell/projects/clientcare/src/app/broker-manager/models/enums/broker-item-type.enum.ts
export enum BrokerItemTypeEnum {
    Brokerage = 1,
    BrokerageAddress = 2,
    BrokerageBankAccount = 3,
    BrokerageBranch = 4,
    BrokerageBrokerConsultant = 5,
    BrokerageCheck = 6,
    BrokerageContact = 7,
    BrokerageFscaLicenseCategory = 8,
    BrokerageNote = 9,
    BrokerageProductOption = 10,
    BrokerageRepresentative = 11,
    BrokerPartnership = 12,
    OrganisationOptionItemValue = 13       
}